Added Meditation Categories
Added the remaining audio along with a switch to allow you to swap between sleep and meditation.

Changed the way the BearerID was saved from storing it in the App.Config to a TXT file labled Bearer, in order to make updating the program less of hassle.

Renamed various functions
Will eventuall restructure the JSONs to idealy be a single entity if possible and then have my API handler take arguments instead of having fifty billion different functions to download the different medias