Tons of changes
Added ability to specify what category gets loaded as what

Added some options to disable/enable deletion of files, and to enable/disable the automated merger of the files if applicable

Added a basic popup that shows the program is downloading and not just frozen.

Hopefully fixed a few outlier crashes. Added a few very basic error messages.