Sleepcasts Are Done! I hope!
Reworked the API used by the program to the latest v3 that Headspace has. From initial network traffic it looks like it may only be used for Sleepcasts? not 100% sure but yolo.
But Headspace does us the great honor of premixing the Voice and Ambience so no need to manually combine. FFMPEG might be useless for this T_T

- Integrated new /content-interface/v3/playable-assets flow
- Added --sleepcast flag for Sleepcast downloads
- Added --date option to override playback date (default: today)
- Added --all-tracks / --mixed-only flags
  - MIXED only is default
  - all-tracks downloads VOICE + AMBIENCE + MIXED
- Downloads now show a progress bar, helpful for large Sleepcast files (~50MB)
- Sleepcast tracks are named after their IDs (e.g., SC-408-MIXED-73320.mp3)